Configure pages / and /home
