+++ frontend/src/pages/Home.tsx
@@ -1,15 +1,45 @@
-
+import { useAuth } from "react-oidc-context";
+import { Navigate } from "react-router-dom";
 
 const Home = () => {
+    const auth = useAuth();
 
-  return(
+    const signOutRedirect = () => {
+        // TODO fazer nao usar direto do .env
+        const clientId = import.meta.env.VITE_COGNITO_CLIENT_ID as string;
+        const logoutUri = import.meta.env.VITE_COGNITO_LOGOUT_URI as string;
+        const cognitoDomain = import.meta.env.VITE_COGNITO_DOMAIN as string;
+        auth.removeUser();
+        window.location.href = `${cognitoDomain}/logout?client_id=${clientId}&logout_uri=${encodeURIComponent(logoutUri)}`;
+    };
 
-        <div>
-            <h1>Welcome to the Home Page</h1>
-            <p>you are logged.</p>
-        </div>
+    if (auth.isLoading) {
+        return <div>Loading...</div>;
+    }
 
-    );
+    if (auth.error) {
+        return <div>Encountering error... {auth.error.message}</div>;
+    }
+
+
+    if (auth.isAuthenticated){
+        return (
+            <div>
+                <h1>Bem vindo {auth.user?.profile.given_name} {auth.user?.profile.family_name}!</h1>
+                <pre> Email: {auth.user?.profile.email} </pre>
+                <pre> Profile: {JSON.stringify(auth.user?.profile, null, 2)}</pre>
+                <pre> ID Token: {auth.user?.id_token} </pre>
+                <pre> Access Token: {auth.user?.access_token} </pre>
+                <pre> Refresh Token: {auth.user?.refresh_token} </pre>
+
+                <button onClick={() => signOutRedirect()}>Sign out</button>
+            </div>
+        );
+    }
+
+    if (!auth.isAuthenticated){
+        return <Navigate to="/" replace />;
+    }
 
 
 };
